feat: moved friends to panels tab
