+++ components/Sidebar/panels/index.tsx
@@ -1,11 +1,11 @@
-import Friends from "./Friends/Friends";
+import Friends from "./Friends";
 import Notifications from "./Notifications";
-
-import type { FirestoreProfile, FirestoreUser } from "@/lib/firestore/types";
 
 const SidebarPanels = ({ isDesktop }: { isDesktop?: boolean }) => (
   <div className={isDesktop ? "hidden md:block" : "md:hidden"}>
+    {/* @ts-expect-error Next 13 handles async components */}
     <Friends />
+    {/* @ts-expect-error Next 13 handles async components */}
     <Notifications />
   </div>
 );
